+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Anvil Level Design Changelog
 
 ## 1.4.2
+- Added option to enable vertex colours on materials (automatically mixes in the colour attribute with the image texture) (colour attribute always added by default)
 - Walk Navigation hold works in vertex paint mode (quick colour is rebound to 'C' by default)
 - Fix regression caused by blender update stopping initial texture apply on default cube and new boxes in object mode